+++ ldso/ldso/hash.c
@@ -94,10 +94,10 @@ struct elf_resolve *_dl_check_hashed_files(char *libname)
  */
 
 struct elf_resolve *_dl_add_elf_hash_table(char *libname, 
-	char *loadaddr, unsigned int *dynamic_info, unsigned int dynamic_addr, 
-	unsigned int dynamic_size)
+	char *loadaddr, unsigned long *dynamic_info, unsigned long dynamic_addr, 
+	unsigned long dynamic_size)
 {
-	unsigned int *hash_addr;
+	unsigned long *hash_addr;
 	struct elf_resolve *tpnt;
 	int i;
 
@@ -123,7 +123,7 @@ struct elf_resolve *_dl_add_elf_hash_table(char *libname,
 	tpnt->libtype = loaded_file;
 
 	if (dynamic_info[DT_HASH] != 0) {
-		hash_addr = (unsigned int *) (dynamic_info[DT_HASH] + loadaddr);
+		hash_addr = (unsigned long *) (dynamic_info[DT_HASH] + loadaddr);
 		tpnt->nbucket = *hash_addr++;
 		tpnt->nchain = *hash_addr++;
 		tpnt->elf_buckets = hash_addr;
@@ -143,7 +143,7 @@ struct elf_resolve *_dl_add_elf_hash_table(char *libname,
  */
 
 char *_dl_find_hash(char *name, struct dyn_elf *rpnt1, 
-	unsigned int instr_addr, struct elf_resolve *f_tpnt, int copyrel)
+	unsigned long instr_addr, struct elf_resolve *f_tpnt, int copyrel)
 {
 	struct elf_resolve *tpnt;
 	int si;
@@ -151,7 +151,7 @@ char *_dl_find_hash(char *name, struct dyn_elf *rpnt1,
 	int pass;
 	char *strtab;
 	Elf32_Sym *symtab;
-	unsigned int elf_hash_number, hn;
+	unsigned long elf_hash_number, hn;
 	char *weak_result;
 	struct elf_resolve *first_def;
 	struct dyn_elf *rpnt, first;
